Add a very basic "PCMCIA" driver
Works with a SCR24x devices and in theory should work with the cm4040
devices which is this modelled after (not tested as I don't have the
hardware).

This does something equivalent to openct's pcmcia_block driver. It's not
in fact PCMCIA specific, but the existing drivers are all for PCMCIA
devices.

The device essentially isn't more than an ordinary character device that speaks
CCID via standard read() and write().